+++ message/rpc.go
@@ -51,53 +51,78 @@ func rpcListen(bind string) {
 	rpc.Accept(l)
 }
 
-// Store offline pravite message interface
-func (r *MessageRPC) Save(m *myrpc.MessageSaveArgs, ret *int) error {
-	if m == nil || m.Msg == nil || m.MsgId < 0 || m.GroupId < 0 {
+// SavePrivate rpc interface save user private message.
+func (r *MessageRPC) SavePrivate(m *myrpc.MessageSavePrivateArgs, ret *int) error {
+	if m == nil || m.Msg == nil || m.MsgId < 0 {
 		return myrpc.ErrParam
 	}
-	if err := UseStorage.Save(m.Key, m.Msg, m.MsgId, m.GroupId, m.Expire); err != nil {
-		glog.Errorf("UseStorage.Save(\"%s\", \"%s\", %d, %d, %d) error(%v)", m.Key, string(m.Msg), m.MsgId, m.GroupId, m.Expire, err)
+	if err := UseStorage.SavePrivate(m.Key, m.Msg, m.MsgId, m.Expire); err != nil {
+		glog.Errorf("UseStorage.SavePrivate(\"%s\", \"%s\", %d, %d) error(%v)", m.Key, string(m.Msg), m.MsgId, m.Expire, err)
 		return err
 	}
-	glog.V(1).Infof("UseStorage.Save(\"%s\", \"%s\", %d, %d, %d) ok", m.Key, string(m.Msg), m.MsgId, m.GroupId, m.Expire)
+	glog.V(1).Infof("UseStorage.SavePrivate(\"%s\", \"%s\", %d, %d) ok", m.Key, string(m.Msg), m.MsgId, m.Expire)
 	return nil
 }
 
-// Store offline public message interface
-func (r *MessageRPC) SavePub(m *myrpc.MessageSavePubArgs, ret *int) error {
-	return nil
-}
-
-// Get offline message interface
-func (r *MessageRPC) Get(m *myrpc.MessageGetArgs, rw *myrpc.MessageGetResp) error {
+// GetPrivate rpc interface get user private message.
+func (r *MessageRPC) GetPrivate(m *myrpc.MessageGetPrivateArgs, rw *myrpc.MessageGetResp) error {
 	if m == nil || m.Key == "" || m.MsgId < 0 {
 		return myrpc.ErrParam
 	}
-	msgs, err := UseStorage.Get(m.Key, m.MsgId)
+	msgs, err := UseStorage.GetPrivate(m.Key, m.MsgId)
 	if err != nil {
-		glog.Errorf("UseStorage.Get(\"%s\", %d) error(%v)", m.Key, m.MsgId, err)
+		glog.Errorf("UseStorage.GetPrivate(\"%s\", %d) error(%v)", m.Key, m.MsgId, err)
 		return err
 	}
 	rw.Msgs = msgs
-	// TODO
-	rw.PubMsgs = nil
-	glog.V(1).Infof("UserStorage.Get(\"%s\", %d) ok", m.Key, m.MsgId)
+	glog.V(1).Infof("UserStorage.GetPrivate(\"%s\", %d) ok", m.Key, m.MsgId)
 	return nil
 }
 
-// Clean offline message interface
-func (r *MessageRPC) Clean(key string, ret *int) error {
+// DelPrivate rpc interface delete user private message.
+func (r *MessageRPC) DelPrivate(key string, ret *int) error {
 	if key == "" {
 		return myrpc.ErrParam
 	}
-	if err := UseStorage.Del(key); err != nil {
-		glog.Errorf("UserStorage.Del(\"%s\") error(%v)", key, err)
+	if err := UseStorage.DelPrivate(key); err != nil {
+		glog.Errorf("UserStorage.DelPrivate(\"%s\") error(%v)", key, err)
 		return err
 	}
-	glog.V(1).Infof("UserStorage.Del(\"%s\") ok", key)
+	glog.V(1).Infof("UserStorage.DelPrivate(\"%s\") ok", key)
 	return nil
 }
+
+/*
+// SavePublish rpc interface save publish message.
+func (r *MessageRPC) SavePublish(m *myrpc.MessageSaveGroupArgs, ret *int) error {
+	return nil
+}
+
+// GetPublish rpc interface get publish message.
+func (r *MessageRPC) GetPublish(m *myrpc.MessageGetGroupArgs, rw *myrpc.MessageGetResp) error {
+	return nil
+}
+
+// DelPublish rpc interface delete publish message.
+func (r *MessageRPC) DelPublish(key string, ret *int) error {
+	return nil
+}
+
+// SaveGroup rpc interface save publish message.
+func (r *MessageRPC) SaveGroup(m *myrpc.MessageSaveGroupArgs, ret *int) error {
+	return nil
+}
+
+// GetPublish rpc interface get publish message.
+func (r *MessageRPC) GetGroup(m *myrpc.MessageGetGroupArgs, rw *myrpc.MessageGetResp) error {
+	return nil
+}
+
+// DelPublish rpc interface delete publish message.
+func (r *MessageRPC) DelGroup(key string, ret *int) error {
+	return nil
+}
+*/
 
 // Server Ping interface
 func (r *MessageRPC) Ping(p int, ret *int) error {