+++ ver/ver.go
@@ -17,5 +17,5 @@
 package ver
 
 const (
-	Version = "1.0.2"
+	Version = "1.0.3"
 )