+++ web/handle_1.0.go
@@ -80,11 +80,10 @@ func GetOfflineMsg(w http.ResponseWriter, r *http.Request) {
 	params := r.URL.Query()
 	key := params.Get("k")
 	midStr := params.Get("m")
-	pmidStr := params.Get("p")
 	callback := params.Get("cb")
 	res := map[string]interface{}{"ret": OK}
 	defer retWrite(w, r, res, callback, time.Now())
-	if key == "" || midStr == "" || pmidStr == "" {
+	if key == "" || midStr == "" {
 		res["ret"] = ParamErr
 		return
 	}
@@ -94,26 +93,20 @@ func GetOfflineMsg(w http.ResponseWriter, r *http.Request) {
 		glog.Errorf("strconv.ParseInt(\"%s\", 10, 64) error(%v)", midStr, err)
 		return
 	}
-	pmid, err := strconv.ParseInt(pmidStr, 10, 64)
-	if err != nil {
-		res["ret"] = ParamErr
-		glog.Errorf("strconv.ParseInt(\"%s\", 10, 64) error(%v)", pmidStr, err)
-		return
-	}
 	// RPC get offline messages
 	reply := &myrpc.MessageGetResp{}
-	args := &myrpc.MessageGetArgs{MsgId: mid, PubMsgId: pmid, Key: key}
+	args := &myrpc.MessageGetPrivateArgs{MsgId: mid, Key: key}
 	client := myrpc.MessageRPC.Get()
 	if client == nil {
 		res["ret"] = InternalErr
 		return
 	}
-	if err := client.Call(myrpc.MessageServiceGet, args, reply); err != nil {
-		glog.Errorf("myrpc.MessageRPC.Call(\"%s\", \"%v\", reply) error(%v)", myrpc.MessageServiceGet, args, err)
+	if err := client.Call(myrpc.MessageServiceGetPrivate, args, reply); err != nil {
+		glog.Errorf("myrpc.MessageRPC.Call(\"%s\", \"%v\", reply) error(%v)", myrpc.MessageServiceGetPrivate, args, err)
 		res["ret"] = InternalErr
 		return
 	}
-	res["data"] = map[string]interface{}{"msgs": reply.Msgs, "pmsgs": reply.PubMsgs}
+	res["data"] = map[string]interface{}{"msgs": reply.Msgs}
 	return
 }